pyproject.toml
[project]
name = "qr-code-tracker"
version = "1.0.0"
description = "QR Code Views Tracker - Generate QR codes that track views and display statistics"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "Flask==2.3.2",
    "Flask-Migrate==4.0.4",
    "flask-sqlalchemy==3.0.3",
    "SQLAlchemy>=2.0.23",
    "waitress==2.1.2",
    "click>=8.0.0",
    "tomli>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
qr-tracker = "src.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
